make single-funding use ZeroconfPublishedFundingTx
